SSL: use certifi if available
Use the certifi package, if it's available. This to address the
case where `ssl` cannot access the system SSL certificates and raises
`certificate verify failed: unable to get local issuer certificate`.

This change also complexify the SSL code, an alternative is to just
ask the users to fix their set-up or disable SSL.